Fix build on amd64

+++ sys/dev/sdhci/sdhci.c
@@ -1991,7 +1991,7 @@ sdhci_cam_request(struct sdhci_slot *slot, union ccb *ccb)
 	if (mmcio->cmd.data != NULL) {
 		if (mmcio->cmd.data->len == 0 || mmcio->cmd.data->flags == 0)
 			panic("data->len = %d, data->flags = %d -- something is b0rked",
-			      mmcio->cmd.data->len, mmcio->cmd.data->flags);
+			      (int)mmcio->cmd.data->len, mmcio->cmd.data->flags);
 	}
 	slot->ccb = ccb;
 	slot->flags = 0;
